requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels-redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
django==6.0
django-js-asset==3.1.2
django-prose-editor==0.22.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
incremental==24.11.0
kombu==5.6.2
msgpack==1.1.2
nh3==0.3.2
packaging==25.0
phonenumbers==9.0.21
pillow==12.1.0
prompt-toolkit==3.0.52
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pyjwt==2.10.1
pyopenssl==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
service-identity==24.2.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.5
stripe==14.1.0
twisted==25.5.0
txaio==25.12.2
typing-extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
zope-interface==8.2
